requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
Werkzeug==3.0.3
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.2
email-validator==2.2.0
